Move TypeVars _A, _B and _C to tensor_dict.py
